1. 채팅 이미지 업로드 기능 추가
2. BitmapTranslator 클래스 생성
2-1. toBitmap(...) 메소드로 String 자료형 인자를 Bitmap으로 변환함
2-2. toBase64(...) 메소드로 Bitmap 객체를 String 문자열로 변환함
2-3. isImage(...) 메소드로 해당 문자열이 이미지인지 구분
2-4. dpToPx() 메소드로 dp 값을 입력하면 픽셀 값으로 변환함
2-5. 위 함수들은 홍석씨의 코드를 참조하여 재작성하였음

3. MessageAdapter 일부 메소드 개선
3-1. onBindViewHolder(...) 메소드의 내부 코드를 함축하였음
3-2. addItem(...) -> addMessage(...) 명칭 변경
3-3. DB 서버에 메시지를 담아 메시지가 저장되도록 하는 기능 구현을 위해 추후 추가 변경 가능성이 다분함

4. RoomLayout 메소드에서 이미지를 불러오며, 이 이미지 기존의 비율에 맞게 알맞은 크기로 리사이징시켜 업로드 하는 기능 구현
4-1. Ratio 클래스를 통해 이미지의 비율을 계산하는 기능을 구현함
4-2. 위 기능들을 활용하여 용량이 큰 사진을 업로드할때 오류가 발생하는 부분을 수정함

+++ app/src/main/java/com/yeungjin/translogic/object/chat/Room.java
@@ -2,10 +2,10 @@ package com.yeungjin.translogic.object.chat;
 
 public class Room {
     public String name;
-    public int number;
+    public long number;
     public String title;
 
-    public Room(String name, int number, String title) {
+    public Room(String name, long number, String title) {
         this.name = name;
         this.number = number;
         this.title = title;
